# client/public/api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: ByteStash API
  version: "1.0.0"
  description: API documentation for ByteStash snippet storage service.
servers:
  - url: /
    description: Local server
components:
  securitySchemes:
    AccessAuth:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: >
        Cloudflare Access JWT. Injected automatically by Cloudflare when the
        request passes an Access policy; browsers send it via the
        CF_Authorization cookie. Not something API clients set manually —
        use an API key instead.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for CLI access
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        ByteStash API key sent as a bearer token (Authorization: Bearer <key>).
        Used by the MCP endpoint, which most remote MCP clients authenticate with.
  schemas:
    Fragment:
      type: object
      properties:
        id:
          type: integer
        file_name:
          type: string
        code:
          type: string
        language:
          type: string
        position:
          type: integer
    Snippet:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        user_id:
          type: integer
        username:
          type: string
        is_public:
          type: integer
          description: 0 or 1
        is_pinned:
          type: integer
          description: 0 or 1
        is_favorite:
          type: integer
          description: 0 or 1
        expiry_date:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/Fragment'
        updated_at:
          type: string
        share_count:
          type: integer
    PaginatedSnippets:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Snippet'
        pagination:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
            hasMore:
              type: boolean
    Share:
      type: object
      properties:
        id:
          type: string
        snippetId:
          type: integer
        requiresAuth:
          type: boolean
        expiresIn:
          type: integer
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
        created_at:
          type: string
    Attachment:
      type: object
      properties:
        id:
          type: integer
        snippet_id:
          type: integer
        file_name:
          type: string
        mime:
          type: string
        size:
          type: integer
        created_at:
          type: string
paths:
  /api/auth/config:
    get:
      summary: Get authentication configuration
      tags:
        - Authentication
      responses:
        '200':
          description: Auth config object
          content:
            application/json:
              schema:
                type: object
                properties:
                  authRequired:
                    type: boolean
                  allowNewAccounts:
                    type: boolean
                  hasUsers:
                    type: boolean
                  disableAccounts:
                    type: boolean
                  disableInternalAccounts:
                    type: boolean
  /api/auth/verify:
    get:
      summary: Resolve the current user from the Cloudflare Access identity
      tags:
        - Authentication
      security:
        - AccessAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
        '401':
          description: Invalid token
  /api/auth/anonymous:
    post:
      summary: Create an anonymous session
      tags:
        - Authentication
      responses:
        '200':
          description: Anonymous session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
        '403':
          description: Anonymous login not allowed
  /api/snippets:
    get:
      summary: Get all snippets for current user (paginated)
      tags:
        - Snippets
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Page size (1-100, default 50)
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of snippets to skip
        - name: search
          in: query
          schema:
            type: string
          description: Filter by title/description (and code when searchCode=true)
        - name: searchCode
          in: query
          schema:
            type: boolean
          description: Include fragment code in the search
        - name: language
          in: query
          schema:
            type: string
          description: Filter by fragment language
        - name: category
          in: query
          schema:
            type: string
          description: Comma-separated list of categories (matches snippets having ALL of them)
        - name: favorites
          in: query
          schema:
            type: boolean
          description: Only return favorited snippets
        - name: pinned
          in: query
          schema:
            type: boolean
          description: Only return pinned snippets
        - name: recycled
          in: query
          schema:
            type: boolean
          description: Only return snippets in the recycle bin
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, alpha-asc, alpha-desc]
            default: newest
          description: Sort order
      responses:
        '200':
          description: Paginated list of snippets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSnippets'
        '401':
          description: Authentication required
    post:
      summary: Create a new snippet
      tags:
        - Snippets
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Snippet'
      responses:
        '201':
          description: Snippet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Authentication required
  /api/snippets/{id}:
    get:
      summary: Get snippet by id
      tags:
        - Snippets
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Snippet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Authentication required
        '404':
          description: Snippet not found
    put:
      summary: Update snippet by id
      tags:
        - Snippets
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Snippet'
      responses:
        '200':
          description: Updated snippet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Authentication required
        '404':
          description: Snippet not found
    delete:
      summary: Delete snippet
      tags:
        - Snippets
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Snippet deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '401':
          description: Authentication required
        '404':
          description: Snippet not found
  /api/snippets/{id}/attachments:
    get:
      summary: List attachments of a snippet
      tags:
        - Attachments
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Attachment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attachment'
        '404':
          description: Snippet not found
    post:
      summary: Upload an attachment (raw body streamed to R2)
      tags:
        - Attachments
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-file-name
          in: header
          required: true
          description: URL-encoded file name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Attachment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '413':
          description: Attachment too large (max 25 MB)
  /api/snippets/{id}/attachments/{attachmentId}:
    get:
      summary: Download an attachment
      tags:
        - Attachments
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Attachment content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Attachment not found
    delete:
      summary: Delete an attachment
      tags:
        - Attachments
      security:
        - AccessAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Attachment deleted
        '404':
          description: Attachment not found
  /api/public/snippets:
    get:
      summary: Get all public snippets
      tags:
        - Public
      responses:
        '200':
          description: List of public snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snippet'
        '500':
          description: Internal server error
  /api/public/snippets/{id}:
    get:
      summary: Get a public snippet
      tags:
        - Public
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A public snippet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '404':
          description: Snippet not found
        '500':
          description: Internal server error
  /api/share:
    post:
      summary: Create share link
      tags:
        - Sharing
      security:
        - AccessAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Share'
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '400':
          description: Invalid snippet ID
        '401':
          description: Authentication required
        '403':
          description: Permission denied
  /api/share/{id}:
    get:
      summary: Retrieve a share
      tags:
        - Sharing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Snippet with share data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Authentication required for protected share
        '404':
          description: Share not found
        '410':
          description: Share has expired
    delete:
      summary: Delete a share
      tags:
        - Sharing
      security:
        - AccessAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          description: Authentication required
        '500':
          description: Failed to delete share
  /api/share/snippet/{snippetId}:
    get:
      summary: Get all shares for a snippet
      tags:
        - Sharing
      security:
        - AccessAuth: []
      parameters:
        - name: snippetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of shares for the snippet
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'
        '401':
          description: Authentication required
  /api/keys:
    get:
      summary: List all API keys for the authenticated user
      tags:
        - API Keys
      security:
        - AccessAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '401':
          description: Authentication required
        '500':
          description: Failed to fetch API keys
    post:
      summary: Create a new API key
      tags:
        - API Keys
      security:
        - AccessAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          description: Name is required
        '401':
          description: Authentication required
        '500':
          description: Failed to create API key
  /api/keys/{id}:
    delete:
      summary: Delete an API key
      tags:
        - API Keys
      security:
        - AccessAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucess:
                    type: boolean
        '401':
          description: Authentication required
        '404':
          description: API key not found
        '500':
          description: Failed to delete API key
  /api/embed/{shareId}:
    get:
      summary: Get a snippet for embedding
      tags:
        - Embed
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
        - name: showTitle
          in: query
          schema:
            type: boolean
        - name: showDescription
          in: query
          schema:
            type: boolean
        - name: fragmentIndex
          in: query
          schema:
            type: integer
      security:
        - AccessAuth: []
      responses:
        '200':
          description: Snippet data for embedding
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  language:
                    type: string
                  fragments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fragment'
                  created_at:
                    type: string
                  updated_at:
                    type: string
        '401':
          description: Authentication required for protected share
        '404':
          description: Snippet not found
        '500':
          description: Internal server error
  /mcp:
    post:
      summary: Remote MCP (Model Context Protocol) endpoint for AI clients
      description: >
        JSON-RPC 2.0 endpoint over the MCP Streamable HTTP transport. Lets MCP-capable
        AI clients (Claude, OpenAI, Perplexity) list, read, create, update and delete the
        authenticated user's snippets. Authenticate with a ByteStash API key sent either as
        `Authorization: Bearer <key>` or the `x-api-key` header. Available tools: list_snippets,
        get_snippet, create_snippet, update_snippet, delete_snippet, list_metadata. See the
        README "MCP (AI assistants)" section for client configuration. The request/response
        bodies are JSON-RPC envelopes, so they are not modelled in detail here.
      tags:
        - MCP
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A JSON-RPC 2.0 request (e.g. initialize, tools/list, tools/call)
      responses:
        '200':
          description: JSON-RPC 2.0 response
        '401':
          description: Missing or invalid API key
        '405':
          description: Method not allowed (use POST for MCP requests)